refactor: remove old out and refactore wishlist
 Deleted the AuthGuard and AuthGuardMiddleware files to streamline authentication handling.
- Updated AuthRequest type to include additional user properties: walletAddress, name, and role.
- Refactored WishlistController to utilize JwtAuthGuard instead of the removed AuthGuard.
- Adjusted user ID handling in controller methods to ensure compatibility with new AuthRequest structure.
- Corrected import paths for Product entity across various files to reflect new module structure.
