+++ src/modules/wishlist/common/types/auth-request.type.ts
@@ -1,9 +1,14 @@
 import { Request } from 'express';
+import { Role } from '../../../auth/entities/role.entity';
 
 export interface AuthRequest extends Request {
   user: {
-    id: string;
+    id: string | number;
+    walletAddress: string;
+    name?: string;
     email?: string;
-    roles?: string[];
+    role: Role[];
+    createdAt?: Date;
+    updatedAt?: Date;
   };
 }
