+++ src/modules/wishlist/controller/wishlist.controller.spec.ts
@@ -1,11 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { WishlistController } from './wishlist.controller';
 import { WishlistService } from '../service/wishlist.service';
-import { AuthGuard } from '../common/guard/auth-guard';
+import { JwtAuthGuard } from '../../shared/guards/jwt-auth.guard';
 import { ConflictException } from '@nestjs/common';
 import { Wishlist } from '../entitities/wishlist.entity';
 import { mockRequest } from '../common/mock/mock-request';
 import { AuthRequest } from '../common/types/auth-request.type';
+import { Role } from '../../auth/entities/role.entity';
 
 describe('WishlistController', () => {
   let controller: WishlistController;
@@ -17,7 +18,7 @@ describe('WishlistController', () => {
       providers: [
         WishlistService,
         {
-          provide: AuthGuard,
+          provide: JwtAuthGuard,
           useValue: {
             canActivate: jest.fn().mockResolvedValue(true),
           },
@@ -33,7 +34,14 @@ describe('WishlistController', () => {
     it('should call the service method to add product', async () => {
       const userId = 'user-id';
       const productId = 'product-id';
-      const req = mockRequest({ user: { id: userId } }) as unknown as AuthRequest;
+      const mockRole = { id: 1, name: 'buyer' } as Role;
+      const req = mockRequest({
+        user: {
+          id: userId,
+          walletAddress: 'test-wallet',
+          role: [mockRole],
+        },
+      }) as unknown as AuthRequest;
 
       jest.spyOn(service, 'addToWishlist').mockResolvedValueOnce(undefined);
 
@@ -44,7 +52,14 @@ describe('WishlistController', () => {
     it('should throw ConflictException when product is already in wishlist', async () => {
       const userId = 'user-id';
       const productId = 'product-id';
-      const req = mockRequest({ user: { id: userId } }) as unknown as AuthRequest;
+      const mockRole = { id: 1, name: 'buyer' } as Role;
+      const req = mockRequest({
+        user: {
+          id: userId,
+          walletAddress: 'test-wallet',
+          role: [mockRole],
+        },
+      }) as unknown as AuthRequest;
 
       jest
         .spyOn(service, 'addToWishlist')
@@ -58,7 +73,14 @@ describe('WishlistController', () => {
     it('should call service method to remove product from wishlist', async () => {
       const userId = 'user-id';
       const productId = 'product-id';
-      const req = mockRequest({ user: { id: userId } }) as unknown as AuthRequest;
+      const mockRole = { id: 1, name: 'buyer' } as Role;
+      const req = mockRequest({
+        user: {
+          id: userId,
+          walletAddress: 'test-wallet',
+          role: [mockRole],
+        },
+      }) as unknown as AuthRequest;
 
       jest.spyOn(service, 'removeFromWishlist').mockResolvedValueOnce(undefined);
 
@@ -70,7 +92,14 @@ describe('WishlistController', () => {
   describe('getWishlist', () => {
     it('should return the user wishlist', async () => {
       const userId = 'user-id';
-      const req = mockRequest({ user: { id: userId } }) as unknown as AuthRequest;
+      const mockRole = { id: 1, name: 'buyer' } as Role;
+      const req = mockRequest({
+        user: {
+          id: userId,
+          walletAddress: 'test-wallet',
+          role: [mockRole],
+        },
+      }) as unknown as AuthRequest;
       const wishlistItems = [new Wishlist()];
 
       jest.spyOn(service, 'getWishlist').mockResolvedValueOnce(wishlistItems);
